Port wieder auf 80 geändert + neue Spalte Tag eingebaut und die Migration gemacht + Eingabefeld name auf Tags geändert + Design etwas geändert

+++ DocumentManagementSystem/API/Controllers/DocumentController.cs
@@ -1,6 +1,10 @@
 using BusinessLayer.Service.Interface;
 using Microsoft.AspNetCore.Mvc;
 using Model;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 
 namespace API.Controllers
 {
@@ -22,7 +26,6 @@ namespace API.Controllers
         public async Task<IActionResult> GetDocuments()
         {
             var documents = await _documentService.GetDocumentsAsync();
-
             return Ok(documents);
         }
 
@@ -52,7 +55,6 @@ namespace API.Controllers
             await _documentService.AddDocumentAsync(newDocumentModel);
 
             return CreatedAtAction(nameof(GetDocumentById), new { documentId = newDocumentModel.Id }, newDocumentModel);
-
         }
 
         //PUT/documents/{documentId}
@@ -61,7 +63,7 @@ namespace API.Controllers
         {
             if (updatedDocumentModel == null || documentId != updatedDocumentModel.Id)
             {
-                return BadRequest("Document data is invalid.");
+                return BadRequest("Invalid document data.");
             }
 
             try
@@ -79,6 +81,13 @@ namespace API.Controllers
         [HttpDelete("{documentId}")]
         public async Task<IActionResult> DeleteDocument(Guid documentId)
         {
+            var document = await _documentService.GetDocumentByIdAsync(documentId);
+
+            if (document == null)
+            {
+                return NotFound($"Document with ID {documentId} not found.");
+            }
+
             await _documentService.DeleteDocumentAsync(documentId);
             return NoContent();
         }
